feature(employment): employment endpoint and some fix

+++ internal/dto/request/photo_profile.go
@@ -1,14 +1,14 @@
 package request
 
 type UpsertPhotoProfileRequest struct {
-	ProfileCode int    `param:"code" validate:"required,numeric"`
+	ProfileCode uint64 `param:"code" validate:"required,numeric"`
 	Base64Image string `json:"base64img" validate:"required"`
 }
 
 type DownloadPhotoProfileRequest struct {
-	ProfileCode int `param:"code" validate:"required,numeric"`
+	ProfileCode uint64 `param:"code" validate:"required,numeric"`
 }
 
 type DeletePhotoProfileRequest struct {
-	ProfileCode int `param:"code" validate:"required,numeric"`
+	ProfileCode uint64 `param:"code" validate:"required,numeric"`
 }
