+++ internal/dto/request/working_experience.go
@@ -1,10 +1,10 @@
 package request
 
 type UpsertWorkingExperienceRequest struct {
-	ProfileCode       int    `param:"code" validate:"required,numeric"`
+	ProfileCode       uint64 `param:"code" validate:"required,numeric"`
 	WorkingExperience string `json:"workingExperience" validate:"required,min=1"`
 }
 
 type GetUserWorkingExperienceRequest struct {
-	ProfileCode int `param:"code" validate:"required,numeric"`
+	ProfileCode uint64 `param:"code" validate:"required,numeric"`
 }
