+++ internal/repository/mysql/profile_photo/profile_photo.go
@@ -9,7 +9,7 @@ import (
 	"gorm.io/gorm"
 )
 
-func (r *Repository) GetUserProfilePhotoByProfileCode(ctx context.Context, profileCode int) (profilePhoto entity.ProfilePhoto, err error) {
+func (r *Repository) GetUserProfilePhotoByProfileCode(ctx context.Context, profileCode uint64) (profilePhoto entity.ProfilePhoto, err error) {
 	result := r.db.Table(profilePhoto.TableName()).Where("profile_code = ?", profileCode).First(&profilePhoto)
 	if err = result.Error; err != nil {
 		if errors.Is(err, gorm.ErrRecordNotFound) {
